Refactor code to separate major functions into individual files
Refactor the code to separate major functions into individual files and add import/export statements.

* Add new files for each major function and constants from `content_script.js`:
  - `emailKeywords.js`, `highlightAnchors.js`, `emailPageCheck.js`, `scholarOneHandleAnchor.js`, `isEndNoteSupportAnchor.js`, `isClarivateEmailList.js`, `handleAnchors.js`, `getEarlierDate.js`, `calculateTimeDifferenceInMinutes.js`, `isValidDateFormat.js`, `convertDateFormat.js`, `hasOpenButNotReopened.js`, `handleCases.js`, `scholarOneStatusColors.js`, `divElementChangerScholarOne.js`, `scholarOneHandleStatus.js`, `generateStyle.js`, `handleStatus.js`, `observer.js`
* Delete `content_script.js`.
* Modify `popup.js` to remove overlapping functionality with `saveSelection.js` and update logic for saving selection.
* Modify `saveSelection.js` to remove overlapping functionality with `popup.js` and update logic for saving selection.
* Update `manifest.json` to reference the new files in the `content_scripts` section.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/clarivatePenang/penangCoeExtensionAI?shareId=XXXX-XXXX-XXXX-XXXX).
