+++ content_script/emailKeywords.js
@@ -1,0 +1,323 @@
+const emailEndNote = "endnote.support@clarivate.com"
+const emailKeywordsEndNote = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'account',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'science',
+  'custserv',
+  'wos', 'WOS',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailWoS = "wosg.support@clarivate.com"
+const emailKeywordsWoS = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'account',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'custserv',
+  'endnote',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailScholarOne = "s1help@clarivate.com";
+const emailKeywordsScholarOne = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'account',
+  'tr.',
+  'queries',
+  'wos', 'WOS',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'custserv',
+  'endnote',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailAccountSupport = "account.support@clarivate.com"
+const emailKeywordsAccountSupport = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'endnote',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'science',
+  'custserv',
+  'wos', 'WOS',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailLifeScience = "lifesciences.support@clarivate.com"
+const emailKeywordsLifeScience = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'endnote',
+  'science',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'science',
+  'custserv',
+  'wos', 'WOS',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailLifeScienceHDS = "DRG.customerservice@clarivate.com"
+const emailKeywordsLifeScienceHDS = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'endnote',
+  'science',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'science',
+  'custserv',
+  'wos', 'WOS',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
+
+const emailLifeSciencePS = ["lsclientservicesdl@clarivate.com", "lifesciences.support@clarivate.com"]
+const emailKeywordsLifeSciencePS = [
+  'ts.',
+  'ts-',
+  'techstreet',
+  'endnote',
+  'science',
+  'tr.',
+  'queries',
+  'jgear',
+  'jstead',
+  'derwent',
+  'customer',
+  'scientific',
+  'proposals',
+  'service',
+  'science',
+  'custserv',
+  'wos', 'WOS',
+  'collections',
+  'invoices',
+  'serion',
+  'services',
+  'compumark',
+  'admin',
+  'contract',
+  'ipsci',
+  'ips',
+  'drg',
+  'dartsip',
+  'hidadataprogram',
+  'cortellis',
+  'compuMark',
+  'account',
+  'billing',
+  'invoice',
+  'certificate',
+  'tax',
+  'support',
+  'askhbi',
+  'cash',
+  'team',
+  'sales',
+  'bis.in', 'bis.mon',
+  'bisqa',
+];
