+++ content_script/generateStyle.js
@@ -1,0 +1,3 @@
+function generateStyle(color) {
+  return `background-color: ${color}; border-radius: 6px; padding: 3px 6px; color: white; font-weight: 500;`;
+}
