+++ content_script/handleAnchors.js
@@ -1,0 +1,15 @@
+function handleAnchors() {
+  const fromFieldDiv = document.getElementsByClassName("standardField uiMenu");
+  for (const fromDiv of fromFieldDiv) {
+    const anchor = fromDiv.querySelector("a.select");
+    if (!isEndNoteSupportAnchor(anchor)) {
+      if (!isClarivateEmailList(anchor)) {
+        highlightAnchorWithSpecificContent(anchor, "red");
+      } else {
+        highlightAnchorWithSpecificContent(anchor, "orange");
+      }
+    } else {
+      unhighlightAnchor(anchor);
+    }
+  }
+}
