+++ content_script/highlightAnchors.js
@@ -1,0 +1,7 @@
+function highlightAnchorWithSpecificContent(anchor, color) {
+  anchor.style.backgroundColor = color;
+}
+
+function unhighlightAnchor(anchor) {
+  anchor.style.backgroundColor = "";
+}
