+++ content_script/observer.js
@@ -1,0 +1,10 @@
+const observer = new MutationObserver(() => {
+  handleAnchors();
+  handleCases();
+  handleStatus();
+});
+
+observer.observe(document, {
+  childList: true,
+  subtree: true,
+});
